bug fix for the clutter
