+++ app/src/main/java/com/example/cicompanion/maps/CampusLocation.kt
@@ -2,6 +2,7 @@ package com.example.cicompanion.maps
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.example.cicompanion.calendar.model.CalendarEvent
 import com.google.android.gms.maps.model.LatLng
 
 enum class LocationType { BUILDING, PARKING, FOOD, AREA, HOUSING, CUSTOM }
@@ -20,3 +21,19 @@ data class CampusLocation(
     val associatedEventId: String? = null,
     val searchKeywords: List<String> = emptyList()
 )
+
+fun isEventAtLocation(event: CalendarEvent, location: CampusLocation): Boolean {
+    if (location.isCustom) {
+        return event.id == location.associatedEventId
+    }
+    
+    val eventLoc = event.location ?: return false
+    
+    // Check main name
+    if (eventLoc.contains(location.name, ignoreCase = true)) return true
+    
+    // Check keywords
+    return location.searchKeywords.any { keyword -> 
+        eventLoc.contains(keyword, ignoreCase = true) 
+    }
+}
