+++ app/src/main/java/com/example/cicompanion/maps/MapScreen.kt
@@ -1,56 +1,29 @@
 package com.example.cicompanion.maps
 
 import android.Manifest
-import android.annotation.SuppressLint
 import android.content.pm.PackageManager
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.*
-import androidx.compose.animation.core.*
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.MenuBook
-import androidx.compose.material.icons.automirrored.filled.Send
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import com.example.cicompanion.calendar.CalendarViewModel
-import com.example.cicompanion.calendar.model.CalendarEvent
 import com.example.cicompanion.social.FirestoreManager
 import com.example.cicompanion.ui.Routes
-import com.example.cicompanion.ui.theme.AppBackground
-import com.example.cicompanion.ui.theme.BrandOrange
 import com.example.cicompanion.ui.theme.CoralRed
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
@@ -58,38 +31,17 @@ import com.google.android.gms.location.Priority
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.LatLngBounds
-import com.google.android.gms.maps.model.MapStyleOptions
 import com.google.firebase.auth.FirebaseAuth
 import com.google.maps.android.compose.*
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
-
-// boundaries for CSUCI
-private val CSUCI_BOUNDS = LatLngBounds(
-    LatLng(34.157, -119.050),
-    LatLng(34.168, -119.035)
-)
-private val CSUCI_CENTER = LatLng(34.16174111410685, -119.04342498111538)
-
-private const val SH_PARKING_DESC = "Student housing parking (only for SH permit holders)"
-private const val GENERAL_PARKING_DESC = "General parking (A, F, E or Visitor permit required)"
-private const val RESTRICTED_PARKING_DESC = "Restricted parking lot (only for R or RV permit holders)"
-
-private val MAP_STYLE_JSON = """
-    [
-      { "featureType": "poi", "stylers": [ { "visibility": "off" } ] },
-      { "featureType": "poi.business", "elementType": "labels", "stylers": [ { "visibility": "off" } ] },
-      { "featureType": "poi.school", "elementType": "labels", "stylers": [ { "visibility": "off" } ] },
-      { "featureType": "poi.park", "elementType": "labels", "stylers": [ { "visibility": "off" } ] }
-    ]
-""".trimIndent()
+import android.annotation.SuppressLint
 
 val campusLocations = listOf(
     CampusLocation(id = "loc_bell_tower", name = "Bell Tower", position = LatLng(34.16138604361421, -119.0432651672823), description = "Center of Campus", type = LocationType.BUILDING, icon = Icons.Default.Business, color = Color(0xFFD32F2F)),
     CampusLocation(id = "loc_bell_tower_east", name = "Bell Tower East", position = LatLng(34.16134665298329, -119.04189180309578), description = "", type = LocationType.BUILDING, icon = Icons.Default.Business, color = Color(0xFFD32F2F), searchKeywords = listOf("BTE")),
     CampusLocation(id = "loc_bell_tower_west", name = "Bell Tower West", position = LatLng(34.16070116130278, -119.04439859472768), description = "", type = LocationType.BUILDING, icon = Icons.Default.Business, color = Color(0xFFD32F2F), searchKeywords = listOf("BTW")),
-    CampusLocation(id = "loc_library", name = "John Spoor Broome Library", position = LatLng(34.16269668565898, -119.04094849136715), description = "Main Library", type = LocationType.BUILDING, icon = Icons.AutoMirrored.Filled.MenuBook, color = Color(0xFF388E3C), searchKeywords = listOf("Broome Library", "Library", "JSB")),
+    CampusLocation(id = "loc_library", name = "John Spoor Broome Library", position = LatLng(34.16269668565898, -119.04094849136715), description = "Main Library", type = LocationType.BUILDING, icon = Icons.Default.MenuBook, color = Color(0xFF388E3C), searchKeywords = listOf("Broome Library", "Library", "JSB")),
     CampusLocation(id = "loc_union", name = "Student Union", position = LatLng(34.1610, -119.0436), description = "Dining and Lounge", type = LocationType.BUILDING, icon = Icons.Default.Groups, color = Color(0xFF388E3C), searchKeywords = listOf("SUB", "Union")),
     CampusLocation(id = "loc_marin", name = "Marin Hall", position = LatLng(34.164528096869034, -119.04507117740494), description = "Faculty Offices for Mathematics and Data Science", type = LocationType.BUILDING, icon = Icons.Default.Business, color = Color(0xFFD32F2F), searchKeywords = listOf("Marin")),
     CampusLocation(id = "loc_shasta", name = "Shasta Hall", position = LatLng(34.164576865185516, -119.04472618829523), description = "Faculty Offices for Computer Science and Engineering", type = LocationType.BUILDING, icon = Icons.Default.School, color = Color(0xFFD32F2F), searchKeywords = listOf("Shasta")),
@@ -156,30 +108,11 @@ val campusLocations = listOf(
     CampusLocation(id = "park_a5", name = "Parking Lot A5", position = LatLng(34.16050864644475, -119.04463491964691), description = GENERAL_PARKING_DESC, type = LocationType.PARKING, icon = Icons.Default.LocalParking, color = Color(0xFF1976D2), searchKeywords = listOf("A5"))
 )
 
-/**
- * Helper function to check if a location matches an event's location string.
- */
-fun isEventAtLocation(event: CalendarEvent, location: CampusLocation): Boolean {
-    if (location.isCustom) {
-        return event.id == location.associatedEventId
-    }
-    
-    val eventLoc = event.location ?: return false
-    
-    // Check main name
-    if (eventLoc.contains(location.name, ignoreCase = true)) return true
-    
-    // Check keywords
-    return location.searchKeywords.any { keyword -> 
-        eventLoc.contains(keyword, ignoreCase = true) 
-    }
-}
-
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun MapScreen(
     navController: NavHostController, 
-    calendarViewModel: CalendarViewModel,
+    calendarViewModel: com.example.cicompanion.calendar.CalendarViewModel,
     mapViewModel: MapViewModel = viewModel(),
     initialLat: Double? = null,
     initialLng: Double? = null,
@@ -519,1205 +452,6 @@ fun MapScreen(
                 )
             }
         }
-    }
-}
-
-@Composable
-fun EventSelectionDialog(
-    customEvents: List<CalendarEvent>,
-    currentEventId: String?,
-    onDismiss: () -> Unit,
-    onConfirm: (String?) -> Unit
-) {
-    var searchQuery by remember { mutableStateOf("") }
-
-    val filteredEvents = remember(searchQuery, customEvents) {
-        customEvents.filter { it.title.contains(searchQuery, ignoreCase = true) }
-    }
-
-    Dialog(onDismissRequest = onDismiss) {
-        Card(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            shape = RoundedCornerShape(24.dp),
-            colors = CardDefaults.cardColors(containerColor = Color.White)
-        ) {
-            @Suppress("DEPRECATION")
-            Column(
-                modifier = Modifier
-                    .padding(20.dp)
-                    .fillMaxWidth()
-            ) {
-                @Suppress("DEPRECATION")
-                Text(
-                    "Link an Event",
-                    style = MaterialTheme.typography.titleLarge,
-                    fontWeight = FontWeight.Bold,
-                    color = CoralRed
-                )
-                Spacer(Modifier.height(16.dp))
-
-                OutlinedTextField(
-                    value = searchQuery,
-                    onValueChange = { searchQuery = it },
-                    placeholder = { Text("Search your events...") },
-                    leadingIcon = { Icon(Icons.Default.Search, contentDescription = null, tint = Color.Gray) },
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(12.dp),
-                    colors = OutlinedTextFieldDefaults.colors(
-                        focusedBorderColor = CoralRed,
-                        unfocusedBorderColor = Color.LightGray
-                    ),
-                    singleLine = true
-                )
-
-                Spacer(Modifier.height(12.dp))
-
-                LazyColumn(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .heightIn(max = 300.dp)
-                ) {
-                    item {
-                        Surface(
-                            onClick = { onConfirm(null) },
-                            modifier = Modifier.fillMaxWidth(),
-                            color = if (currentEventId == null) CoralRed.copy(alpha = 0.1f) else Color.Transparent,
-                            shape = RoundedCornerShape(8.dp)
-                        ) {
-                            Row(
-                                modifier = Modifier.padding(12.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Icon(Icons.Default.LinkOff, contentDescription = null, tint = Color.Gray)
-                                Spacer(Modifier.width(12.dp))
-                                Text("None (Unlink Event)", fontWeight = FontWeight.Medium)
-                            }
-                        }
-                        HorizontalDivider(modifier = Modifier.padding(vertical = 4.dp), color = Color.LightGray.copy(alpha = 0.3f))
-                    }
-
-                    if (filteredEvents.isEmpty()) {
-                        item {
-                            Box(Modifier.fillMaxWidth().padding(24.dp), contentAlignment = Alignment.Center) {
-                                Text("No events found.", color = Color.Gray)
-                            }
-                        }
-                    } else {
-                        items(filteredEvents) { event ->
-                            val isSelected = event.id == currentEventId
-                            Surface(
-                                onClick = { onConfirm(event.id) },
-                                modifier = Modifier.fillMaxWidth(),
-                                color = if (isSelected) CoralRed.copy(alpha = 0.1f) else Color.Transparent,
-                                shape = RoundedCornerShape(8.dp)
-                            ) {
-                                Row(
-                                    modifier = Modifier.padding(12.dp),
-                                    verticalAlignment = Alignment.CenterVertically
-                                ) {
-                                    Icon(Icons.Default.Event, contentDescription = null, tint = if (isSelected) CoralRed else Color.Gray)
-                                    Spacer(Modifier.width(12.dp))
-                                    Column {
-                                        Text(event.title, fontWeight = FontWeight.Bold, color = if (isSelected) CoralRed else Color.Black)
-                                        Text(event.timeLabel(), style = MaterialTheme.typography.bodySmall, color = Color.Gray)
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                Spacer(Modifier.height(20.dp))
-                Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.End) {
-                    @Suppress("DEPRECATION")
-                    TextButton(onClick = onDismiss) {
-                        @Suppress("DEPRECATION")
-                        Text("Cancel", color = Color.Gray)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun PinCreationDialog(
-    customEvents: List<CalendarEvent>,
-    editingPin: CustomPin? = null,
-    onDismiss: () -> Unit,
-    onConfirm: (String, String, Color, String?) -> Unit
-) {
-    var name by remember { mutableStateOf(editingPin?.name ?: "") }
-    var description by remember { mutableStateOf(editingPin?.description ?: "") }
-    var selectedColor by remember { mutableStateOf(editingPin?.color ?: CoralRed) }
-    var selectedEventId by remember { mutableStateOf<String?>(editingPin?.associatedEventId) }
-    var showEventPicker by remember { mutableStateOf(false) }
-
-    val colors = listOf(CoralRed, Color(0xFF9C27B0), Color(0xFF2196F3), Color(0xFF4CAF50), Color(0xFFFF9800))
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text(if (editingPin != null) "Edit Custom Pin" else "Create Custom Pin", fontWeight = FontWeight.Bold, color = CoralRed) },
-        text = {
-            Column(modifier = Modifier.verticalScroll(rememberScrollState())) {
-                OutlinedTextField(
-                    value = name,
-                    onValueChange = { name = it },
-                    label = { Text("Name") },
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(12.dp),
-                    colors = OutlinedTextFieldDefaults.colors(focusedBorderColor = CoralRed, focusedLabelColor = CoralRed)
-                )
-                Spacer(Modifier.height(8.dp))
-                OutlinedTextField(
-                    value = description,
-                    onValueChange = { description = it },
-                    label = { Text("Description") },
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(12.dp),
-                    colors = OutlinedTextFieldDefaults.colors(focusedBorderColor = CoralRed, focusedLabelColor = CoralRed)
-                )
-                Spacer(Modifier.height(16.dp))
-
-                Text("Pick a Color", style = MaterialTheme.typography.labelLarge)
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp), modifier = Modifier.padding(top = 8.dp)) {
-                    colors.forEach { color ->
-                        Box(
-                            modifier = Modifier
-                                .size(32.dp)
-                                .background(color, CircleShape)
-                                .border(if (selectedColor == color) 2.dp else 0.dp, Color.Black, CircleShape)
-                                .clickable { selectedColor = color }
-                        )
-                    }
-                }
-
-                Spacer(Modifier.height(24.dp))
-                Text("Linked Event", style = MaterialTheme.typography.labelLarge)
-                Spacer(Modifier.height(8.dp))
-
-                val selectedEvent = customEvents.find { it.id == selectedEventId }
-                Surface(
-                    onClick = { showEventPicker = true },
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(12.dp),
-                    border = BorderStroke(1.dp, Color.LightGray),
-                    color = Color(0xFFF8F8F8)
-                ) {
-                    Row(
-                        modifier = Modifier.padding(12.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            imageVector = if (selectedEvent != null) Icons.Default.Event else Icons.Default.LinkOff,
-                            contentDescription = null,
-                            tint = if (selectedEvent != null) CoralRed else Color.Gray
-                        )
-                        Spacer(Modifier.width(12.dp))
-                        Text(
-                            text = selectedEvent?.title ?: "No event linked (Tap to select)",
-                            color = if (selectedEvent != null) Color.Black else Color.Gray,
-                            modifier = Modifier.weight(1f)
-                        )
-                        Icon(Icons.Default.ChevronRight, contentDescription = null, tint = Color.Gray)
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            Button(
-                onClick = { onConfirm(name, description, selectedColor, selectedEventId) },
-                enabled = name.isNotBlank(),
-                colors = ButtonDefaults.buttonColors(containerColor = CoralRed),
-                shape = RoundedCornerShape(12.dp)
-            ) {
-                @Suppress("DEPRECATION")
-                Text(if (editingPin != null) "Save Changes" else "Confirm")
-            }
-        },
-        dismissButton = {
-            @Suppress("DEPRECATION")
-            TextButton(onClick = onDismiss) { @Suppress("DEPRECATION") Text("Cancel", color = Color.Gray) }
-        }
-    )
-
-    if (showEventPicker) {
-        EventSelectionDialog(
-            customEvents = customEvents,
-            currentEventId = selectedEventId,
-            onDismiss = { showEventPicker = false },
-            onConfirm = {
-                selectedEventId = it
-                showEventPicker = false
-            }
-        )
-    }
-}
-
-@Composable
-fun LocationDetailsContent(
-    location: CampusLocation, 
-    events: List<CalendarEvent>,
-    onGoToEvent: (CalendarEvent) -> Unit,
-    onDeletePin: () -> Unit,
-    onTogglePin: () -> Unit,
-    onSendMessage: () -> Unit,
-    onAssociateEvent: () -> Unit,
-    onEditPin: () -> Unit,
-    onSaveSharedPin: () -> Unit = {}
-) {
-    val isTemp = location.id == "shared_temp"
-    val detailsContentModifier = Modifier
-        .fillMaxWidth()
-        .padding(horizontal = 24.dp)
-
-    @Suppress("DEPRECATION")
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .verticalScroll(rememberScrollState())
-            .padding(top = 24.dp, bottom = 32.dp)
-    ) {
-        Row( modifier = detailsContentModifier,
-            verticalAlignment = Alignment.CenterVertically) {
-            Box(
-                modifier = Modifier
-                    .size(56.dp)
-                    .background(location.color.copy(alpha = 0.1f), CircleShape),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(location.icon, contentDescription = null, tint = location.color, modifier = Modifier.size(32.dp))
-            }
-            Spacer(Modifier.width(16.dp))
-            Column(modifier = Modifier.weight(1f)) {
-                @Suppress("DEPRECATION")
-                Text(
-                    text = location.name,
-                    style = MaterialTheme.typography.headlineSmall,
-                    fontWeight = FontWeight.Bold,
-                    color = Color.Black
-                )
-                @Suppress("DEPRECATION")
-                Text(
-                    text = if (isTemp) "Shared Temporary Pin" else location.type.name.lowercase().replaceFirstChar { it.uppercase() },
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = Color.Black
-                )
-
-                Spacer(modifier = Modifier.height(2.dp))
-                @Suppress("DEPRECATION")
-                Text(
-                    text = if (isTemp) "Clicking empty space on the map or another point will remove this pin." else location.type.name.lowercase().replaceFirstChar { it.uppercase() },
-                    style = MaterialTheme.typography.bodySmall,
-                    color = Color.Gray
-                )
-            }
-            if (!isTemp) {
-                Row {
-//                    if (location.isCustom) {
-////                        IconButton(onClick = onEditPin) {
-////                            Icon(Icons.Default.Edit, contentDescription = "Edit Pin", tint = Color.Gray)
-//                        }
-                    }
-                if (location.isCustom) {
-                    IconButton(onClick = onEditPin) {
-                        Icon(Icons.Default.Edit, contentDescription = "Edit Pin", tint = Color.Gray)
-                        }
-                    }
-                }
-            }
-        }
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-    if (location.description.isNotEmpty()) {
-        Column(modifier = detailsContentModifier) {
-            @Suppress("DEPRECATION")
-            Text(
-                text = "About",
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                color = Color.Black
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            @Suppress("DEPRECATION")
-            Text(
-                text = location.description,
-                style = MaterialTheme.typography.bodyLarge,
-                color = Color.DarkGray,
-                lineHeight = 24.sp
-            )
-        }
-
-        Spacer(modifier = Modifier.height(24.dp))
-    }
-
-        Row(
-            modifier = detailsContentModifier,
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            if (isTemp) {
-                Button(
-                    onClick = onSaveSharedPin,
-                    colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF4CAF50)),
-                    modifier = Modifier.weight(1f),
-                    shape = RoundedCornerShape(12.dp)
-                ) {
-                    Icon(Icons.Default.Save, contentDescription = null, modifier = Modifier.size(18.dp))
-                    Spacer(Modifier.width(8.dp))
-                    @Suppress("DEPRECATION")
-                    Text("Save to My Map")
-                }
-            } else {
-                if (location.isCustom) {
-                    Button(
-                        onClick = onDeletePin,
-                        colors = ButtonDefaults.buttonColors(containerColor = Color.Red.copy(alpha = 0.1f), contentColor = Color.Red),
-                        modifier = Modifier.weight(1f),
-                        shape = RoundedCornerShape(12.dp)
-                    ) {
-                        Icon(Icons.Default.Delete, contentDescription = null, modifier = Modifier.size(18.dp))
-                        Spacer(Modifier.width(4.dp))
-                        @Suppress("DEPRECATION")
-                        Text("Delete")
-                    }
-                    Button(
-                        onClick = onSendMessage,
-                        colors = ButtonDefaults.buttonColors(containerColor = CoralRed.copy(alpha = 0.1f), contentColor = CoralRed),
-                        modifier = Modifier.weight(1f),
-                        shape = RoundedCornerShape(12.dp)
-                    ) {
-                        Icon(Icons.AutoMirrored.Filled.Send, contentDescription = null, modifier = Modifier.size(18.dp))
-                        Spacer(Modifier.width(4.dp))
-                        @Suppress("DEPRECATION")
-                        Text("Share")
-                    }
-                }
-            }
-        }
-
-        if (location.isCustom && !isTemp) {
-            Spacer(modifier = Modifier.height(8.dp))
-            Button(
-                onClick = onAssociateEvent,
-                colors = ButtonDefaults.buttonColors(containerColor = CoralRed),
-                modifier = detailsContentModifier,
-                shape = RoundedCornerShape(12.dp)
-            ) {
-                Icon(Icons.Default.Link, contentDescription = null, modifier = Modifier.size(18.dp))
-                Spacer(Modifier.width(8.dp))
-                @Suppress("DEPRECATION")
-                Text(if (location.associatedEventId != null) "Change Linked Event" else "Link an Event")
-            }
-        }
-        
-        Spacer(modifier = Modifier.height(16.dp))
-
-        if (events.isNotEmpty()) {
-            val visibleEvents = if (isTemp && location.associatedEventId != null) {
-                // If it's a shared pin with an associated event, only show if user is already a member
-                events.filter { it.calendarId == "custom" && it.ownerId != FirebaseAuth.getInstance().currentUser?.uid }
-            } else {
-                events
-            }
-
-            if (visibleEvents.isNotEmpty()) {
-
-                Column(modifier = detailsContentModifier) {
-                    @Suppress("DEPRECATION")
-                    Text(
-                        text = if (location.isCustom && location.associatedEventId != null) {
-                            "Linked Event"
-                        } else {
-                            "Events at this Location"
-                        },
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Bold,
-                        color = Color.Black
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(12.dp))
-
-                visibleEvents.forEach { event ->
-                    Box(modifier = detailsContentModifier) {
-                        MapEventItem(
-                            event = event,
-                            onMoreClick = { onGoToEvent(event) }
-                        )
-                    }
-
-                    Spacer(modifier = Modifier.height(8.dp))
-                }
-            }
-        } else {
-            @Suppress("DEPRECATION")
-            Text(
-                text = "No upcoming events here.",
-                style = MaterialTheme.typography.bodyMedium,
-                color = Color.Gray,
-                modifier = detailsContentModifier.padding(vertical = 8.dp)
-            )
-        }
-
-    Spacer(modifier = Modifier.height(32.dp))
-}
-
-
-
-@Composable
-fun MapEventItem(event: CalendarEvent, onMoreClick: () -> Unit) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        shape = RoundedCornerShape(12.dp),
-        colors = CardDefaults.cardColors(containerColor = Color(0xFFF8F8F8)),
-        border = BorderStroke(1.dp, Color.LightGray.copy(alpha = 0.5f))
-    ) {
-        Row(
-            modifier = Modifier.padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(40.dp)
-                    .background(
-                        when {
-                            event.isPinned -> Color(0xFF9C27B0).copy(alpha = 0.1f)
-                            /* event.isBookmarked -> Color(0xFFFFC107).copy(alpha = 0.1f) */
-                            else -> BrandOrange.copy(alpha = 0.1f)
-                        },
-                        CircleShape
-                    ),
-                contentAlignment = Alignment.Center
-            ) {
-                @Suppress("DEPRECATION")
-                Icon(
-                    imageVector = when {
-                        event.isPinned -> Icons.Default.PushPin
-                        /* event.isBookmarked -> Icons.Default.Bookmark */
-                        else -> Icons.Default.Event
-                    },
-                    contentDescription = null,
-                    tint = when {
-                        event.isPinned -> Color(0xFF9C27B0)
-                        /* event.isBookmarked -> Color(0xFFFFC107) */
-                        else -> BrandOrange
-                    },
-                    modifier = Modifier.size(20.dp)
-                )
-            }
-            Spacer(Modifier.width(12.dp))
-            Column(modifier = Modifier.weight(1f)) {
-                @Suppress("DEPRECATION")
-                Text(
-                    text = event.title,
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 14.sp,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                @Suppress("DEPRECATION")
-                Text(
-                    text = event.timeLabel(),
-                    fontSize = 12.sp,
-                    color = Color.Gray
-                )
-            }
-            
-            // calendar button
-            Box(
-                modifier = Modifier
-                    .size(36.dp)
-                    .clip(CircleShape)
-                    .background(CoralRed.copy(alpha = 0.1f))
-                    .clickable(onClick = onMoreClick),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(
-                    imageVector = Icons.Default.CalendarMonth,
-                    contentDescription = "Go to Calendar",
-                    tint = CoralRed,
-                    modifier = Modifier.size(20.dp)
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun MapTopControls(
-    searchQuery: String,
-    onSearchChange: (String) -> Unit,
-    showSearchResults: Boolean,
-    onClearSearch: () -> Unit,
-    filteredLocations: List<CampusLocation>,
-    onResultClick: (CampusLocation) -> Unit,
-    filterType: LocationType?,
-    onFilterClick: (LocationType?) -> Unit
-) {
-    Surface(
-        shadowElevation = 8.dp,
-        color = Color.White
-    ) {
-        Column(modifier = Modifier.padding(bottom = 12.dp)) {
-            OutlinedTextField(
-                value = searchQuery,
-                onValueChange = onSearchChange,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp, vertical = 12.dp),
-                placeholder = { Text("Search campus...") },
-                leadingIcon = { Icon(Icons.Default.Search, contentDescription = null, tint = Color.Gray) },
-                trailingIcon = {
-                    if (searchQuery.isNotEmpty()) {
-                        IconButton(onClick = onClearSearch) {
-                            Icon(Icons.Default.Close, contentDescription = null, tint = Color.Gray)
-                        }
-                    }
-                },
-                singleLine = true,
-                shape = RoundedCornerShape(16.dp),
-                colors = OutlinedTextFieldDefaults.colors(
-                    unfocusedBorderColor = Color.LightGray.copy(alpha = 0.3f),
-                    focusedBorderColor = CoralRed,
-                    focusedContainerColor = Color(0xFFF8F8F8),
-                    unfocusedContainerColor = Color(0xFFF8F8F8)
-                )
-            )
-
-            if (showSearchResults && filteredLocations.isNotEmpty()) {
-                Card(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 16.dp, vertical = 4.dp)
-                        .heightIn(max = 200.dp),
-                    shape = RoundedCornerShape(12.dp),
-                    elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
-                    colors = CardDefaults.cardColors(containerColor = Color.White)
-                ) {
-                    LazyColumn {
-                        items(filteredLocations) { location ->
-                            ListItem(
-                                headlineContent = { Text(location.name, fontWeight = FontWeight.Medium) },
-                                supportingContent = { Text(location.type.name, fontSize = 12.sp, color = Color.Gray) },
-                                leadingContent = { Icon(location.icon, contentDescription = null, tint = location.color, modifier = Modifier.size(24.dp)) },
-                                modifier = Modifier.clickable { onResultClick(location) }
-                            )
-                        }
-                    }
-                }
-            }
-
-            CategoryFilterRow(selectedType = filterType, onFilterClick = onFilterClick)
-        }
-    }
-}
-
-@Composable
-fun CategoryFilterRow(selectedType: LocationType?, onFilterClick: (LocationType?) -> Unit) {
-    LazyRow(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(10.dp, Alignment.CenterHorizontally),
-        contentPadding = PaddingValues(horizontal = 16.dp)
-    ) {
-        item {
-            CustomFilterChip(
-                selected = selectedType == null,
-                onClick = { onFilterClick(null) },
-                label = "All",
-                icon = null
-            )
-        }
-        LocationType.entries.forEach { type ->
-            item {
-                val icon = when (type) {
-                    LocationType.BUILDING -> Icons.Default.Business
-                    LocationType.FOOD -> Icons.Default.Restaurant
-                    LocationType.AREA -> Icons.Default.People
-                    LocationType.HOUSING -> Icons.Default.Home
-                    LocationType.PARKING -> Icons.Default.LocalParking
-                    LocationType.CUSTOM -> Icons.Default.PushPin
-                }
-                CustomFilterChip(
-                    selected = selectedType == type,
-                    onClick = { onFilterClick(type) },
-                    label = type.name.lowercase().replaceFirstChar { it.uppercase() },
-                    icon = icon
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun CustomFilterChip(
-    selected: Boolean,
-    onClick: () -> Unit,
-    label: String,
-    icon: androidx.compose.ui.graphics.vector.ImageVector?
-) {
-    Surface(
-        onClick = onClick,
-        color = if (selected) CoralRed else Color.White,
-        shape = RoundedCornerShape(12.dp),
-        border = if (selected) null else BorderStroke(1.dp, Color.LightGray.copy(alpha = 0.5f)),
-        modifier = Modifier.height(36.dp)
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.padding(horizontal = 12.dp)
-        ) {
-            if (icon != null) {
-                Icon(
-                    imageVector = icon,
-                    contentDescription = null,
-                    modifier = Modifier.size(16.dp),
-                    tint = if (selected) Color.White else Color.DarkGray
-                )
-                Spacer(Modifier.width(6.dp))
-            }
-            @Suppress("DEPRECATION")
-            Text(
-                text = label,
-                style = MaterialTheme.typography.labelLarge,
-                fontWeight = if (selected) FontWeight.Bold else FontWeight.Normal,
-                color = if (selected) Color.White else Color.DarkGray,
-                fontSize = 13.sp
-            )
-        }
-    }
-}
-
-@Composable
-fun MapContent(
-    cameraPositionState: CameraPositionState,
-    hasLocationPermission: Boolean,
-    userLocation: LatLng?,
-    selectedLocation: CampusLocation?,
-    onLocationClick: (CampusLocation) -> Unit,
-    onMapClick: (LatLng) -> Unit,
-    displayLocations: List<CampusLocation>,
-    events: List<CalendarEvent>,
-    isPinMode: Boolean = false,
-    tempPinLocation: LatLng? = null
-) {
-    val uiSettings = remember(isPinMode) {
-        MapUiSettings(
-            myLocationButtonEnabled = false,
-            zoomControlsEnabled = !isPinMode,
-            compassEnabled = true,
-            mapToolbarEnabled = false
-        )
-    }
-
-    val mapProperties = remember(hasLocationPermission, isPinMode) {
-        MapProperties(
-            isMyLocationEnabled = false,
-            latLngBoundsForCameraTarget = CSUCI_BOUNDS,
-            minZoomPreference = 15f,
-            mapStyleOptions = MapStyleOptions(MAP_STYLE_JSON)
-        )
-    }
-
-    GoogleMap(
-        modifier = Modifier.fillMaxSize(),
-        cameraPositionState = cameraPositionState,
-        properties = mapProperties,
-        uiSettings = uiSettings,
-        onMapClick = onMapClick
-    ) {
-        if (userLocation != null) UserLocationMarker(userLocation)
-
-        if (isPinMode) {
-            tempPinLocation?.let {
-                MarkerComposable(
-                    state = rememberMarkerState(position = it),
-                    anchor = Offset(0.5f, 1.0f)
-                ) {
-                    CustomPinMarkerIcon(
-                        color = CoralRed,
-                        isPinned = false,
-                        eventCount = 0,
-                        isEditing = true
-                    )
-                }
-            }
-        } else {
-            // Draw shared temporary pin if exists
-            selectedLocation?.let { location ->
-                if (location.id == "shared_temp") {
-                    MarkerComposable(
-                        state = rememberMarkerState(position = location.position),
-                        zIndex = 100f,
-                        anchor = Offset(0.5f, 1.0f),
-                        onClick = {
-                            onLocationClick(location)
-                            true
-                        }
-                    ) {
-                        SelectedPointerIcon(location, 0, false, false)
-                    }
-                }
-            }
-
-            displayLocations.forEach { location ->
-                val isSelected = selectedLocation?.id == location.id
-                val locationEvents = events.filter {
-                    (it.calendarId == "custom" || it.isPinned) && isEventAtLocation(it, location)
-                }
-                val eventCount = locationEvents.size
-                val hasPinnedEvent = locationEvents.any { it.isPinned }
-                val hasBookmarkedEvent = false // locationEvents.any { it.isBookmarked }
-
-                key(location.id, location.name, isSelected, eventCount, hasPinnedEvent, hasBookmarkedEvent, location.isCustom, location.isPinned) {
-                    MarkerComposable(
-                        state = rememberMarkerState(position = location.position),
-                        zIndex = if (isSelected) 100f else 1f,
-                        anchor = if (isSelected || location.isCustom) Offset(0.5f, 1.0f) else Offset(0.5f, 0.5f),
-                        onClick = {
-                            onLocationClick(location)
-                            true
-                        }
-                    ) {
-                        if (isSelected) {
-                            SelectedPointerIcon(location, eventCount, hasPinnedEvent, hasBookmarkedEvent)
-                        } else {
-                            if (location.isCustom) {
-                                CustomPinMarkerIcon(location.color, location.isPinned, eventCount)
-                            } else {
-                                LandmarkIcon(
-                                    icon = location.icon,
-                                    color = location.color,
-                                    eventCount = eventCount,
-                                    hasPinnedEvent = hasPinnedEvent,
-                                    hasBookmarkedEvent = hasBookmarkedEvent,
-                                    isCustom = location.isCustom,
-                                    isPinned = location.isPinned
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun CustomPinMarkerIcon(color: Color, isPinned: Boolean, eventCount: Int, isEditing: Boolean = false) {
-    val infiniteTransition = rememberInfiniteTransition(label = "markerPulse")
-    val pulseScale by infiniteTransition.animateFloat(
-        initialValue = 1f,
-        targetValue = if (isEditing) 1.6f else 1.3f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(1200, easing = FastOutSlowInEasing),
-            repeatMode = RepeatMode.Restart
-        ),
-        label = "pulse"
-    )
-    val pulseAlpha by infiniteTransition.animateFloat(
-        initialValue = 0.5f,
-        targetValue = 0f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(1200, easing = FastOutSlowInEasing),
-            repeatMode = RepeatMode.Restart
-        ),
-        label = "pulseAlpha"
-    )
-
-    Box(contentAlignment = Alignment.Center, modifier = Modifier.size(52.dp)) {
-        // Pulsing background ring
-        Box(
-            modifier = Modifier
-                .size(40.dp)
-                .graphicsLayer(scaleX = pulseScale, scaleY = pulseScale)
-                .background(color.copy(alpha = pulseAlpha), CircleShape)
-        )
-
-        // Pin body (Teardrop shape)
-        Box(
-            modifier = Modifier
-                .size(34.dp)
-                .graphicsLayer(rotationZ = 45f)
-                .shadow(elevation = 6.dp, shape = RoundedCornerShape(topStart = 17.dp, topEnd = 17.dp, bottomStart = 17.dp, bottomEnd = 2.dp))
-                .background(color, RoundedCornerShape(topStart = 17.dp, topEnd = 17.dp, bottomStart = 17.dp, bottomEnd = 2.dp))
-                .border(
-                    width = 2.dp,
-                    color = if (isPinned) Color(0xFF9C27B0) else Color.White,
-                    shape = RoundedCornerShape(topStart = 17.dp, topEnd = 17.dp, bottomStart = 17.dp, bottomEnd = 2.dp)
-                )
-        )
-
-        // Inner white circle to make it pop
-        Box(
-            modifier = Modifier
-                .size(20.dp)
-                .background(Color.White, CircleShape)
-                .shadow(2.dp, CircleShape),
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                imageVector = Icons.Default.PushPin,
-                contentDescription = null,
-                tint = color,
-                modifier = Modifier.size(14.dp)
-            )
-        }
-
-        // Badge for associated events
-        if (eventCount > 0) {
-            Box(
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .offset(x = (-4).dp, y = 4.dp)
-                    .size(18.dp)
-                    .background(Color.Red, CircleShape)
-                    .border(1.5.dp, Color.White, CircleShape),
-                contentAlignment = Alignment.Center
-            ) {
-//                @Suppress("DEPRECATION")
-//                Text(
-//                    text = eventCount.toString(),
-//                    color = Color.White,
-//                    fontSize = 10.sp,
-//                    fontWeight = FontWeight.Bold
-//                )
-            }
-        }
-    }
-}
-
-@Composable
-fun SelectedPointerIcon(
-    location: CampusLocation, 
-    eventCount: Int = 0, 
-    hasPinnedEvent: Boolean = false,
-    @Suppress("UNUSED_PARAMETER") hasBookmarkedEvent: Boolean = false
-) {
-    val infiniteTransition = rememberInfiniteTransition(label = "markerBounce")
-    val bounce by infiniteTransition.animateFloat(
-        initialValue = 0f,
-        targetValue = -12f, 
-        animationSpec = infiniteRepeatable(
-            animation = tween(600, easing = FastOutSlowInEasing),
-            repeatMode = RepeatMode.Reverse
-        ),
-        label = "bounce"
-    )
-
-    Box(
-        contentAlignment = Alignment.TopCenter,
-        modifier = Modifier
-            .size(70.dp) 
-            .graphicsLayer(translationY = bounce)
-    ) {
-        Icon(
-            imageVector = Icons.Default.LocationOn,
-            contentDescription = null,
-            tint = location.color,
-            modifier = Modifier.fillMaxSize().shadow(4.dp, CircleShape)
-        )
-        Surface(
-            shape = CircleShape,
-            color = Color.White,
-            modifier = Modifier.padding(top = 8.dp).size(32.dp),
-            border = BorderStroke(2.5.dp, when {
-                location.isPinned -> Color(0xFF9C27B0) // Purple for pinned custom pins
-                hasPinnedEvent -> Color(0xFF9C27B0) // Purple for pinned
-                eventCount > 0 -> BrandOrange // Orange for regular events
-                else -> location.color
-            }),
-            shadowElevation = 4.dp
-        ) {
-            Box(contentAlignment = Alignment.Center) {
-                Icon(
-                    imageVector = if (location.isCustom) Icons.Default.PushPin else location.icon,
-                    contentDescription = null,
-                    tint = location.color,
-                    modifier = Modifier.size(20.dp)
-                )
-            }
-        }
-
-        if (eventCount > 0) {
-            Box(
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .offset(x = (-8).dp, y = 8.dp)
-                    .size(18.dp)
-                    .background(Color.Red, CircleShape)
-                    .border(1.5.dp, Color.White, CircleShape),
-                contentAlignment = Alignment.Center
-            ) {
-//                @Suppress("DEPRECATION")
-//                Text(
-//                    text = eventCount.toString(),
-//                    color = Color.White,
-//                    fontSize = 10.sp,
-//                    fontWeight = FontWeight.Bold
-//                )
-            }
-        }
-    }
-}
-
-@Composable
-fun LandmarkIcon(
-    icon: androidx.compose.ui.graphics.vector.ImageVector, 
-    color: Color, 
-    eventCount: Int = 0, 
-    hasPinnedEvent: Boolean = false,
-    @Suppress("UNUSED_PARAMETER") hasBookmarkedEvent: Boolean = false,
-    isCustom: Boolean = false,
-    isPinned: Boolean = false
-) {
-    Box(contentAlignment = Alignment.Center, modifier = Modifier.size(44.dp)) {
-        Box(
-            modifier = Modifier
-                .size(34.dp)
-                .background(color, CircleShape)
-                .border(2.5.dp, when {
-                    isPinned -> Color(0xFF9C27B0) // Purple border for pinned
-                    hasPinnedEvent -> Color(0xFF9C27B0) // Purple for pinned
-                    eventCount > 0 -> BrandOrange // Orange for regular events
-                    else -> Color.White
-                }, CircleShape)
-                .shadow(4.dp, CircleShape)
-                .padding(7.dp),
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                imageVector = if (isCustom) Icons.Default.PushPin else icon,
-                contentDescription = null, 
-                tint = Color.White, 
-                modifier = Modifier.size(18.dp)
-            )
-        }
-
-        if (eventCount > 0) {
-            Box(
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .offset(x = (-2).dp, y = 2.dp)
-                    .size(16.dp)
-                    .background(Color.Red, CircleShape)
-                    .border(1.5.dp, Color.White, CircleShape),
-                contentAlignment = Alignment.Center
-            ) {
-//                @Suppress("DEPRECATION")
-//                Text(
-//                    text = eventCount.toString(),
-//                    color = Color.White,
-//                    fontSize = 9.sp,
-//                    fontWeight = FontWeight.Bold
-//                )
-            }
-        }
-    }
-}
-
-@Composable
-fun LocationInfoCard(location: CampusLocation, onClose: () -> Unit, onDetailsClick: () -> Unit) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth(0.70f)
-            .padding(bottom = 36.dp),
-        shape = RoundedCornerShape(20.dp),
-        colors = CardDefaults.cardColors(containerColor = Color.White),
-        elevation = CardDefaults.cardElevation(defaultElevation = 12.dp)
-    ) {
-        Column(modifier = Modifier.padding(16.dp)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Box(
-                    modifier = Modifier.size(44.dp).background(location.color.copy(alpha = 0.1f), CircleShape),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Icon(
-                       imageVector = if (location.isCustom) Icons.Default.PushPin else location.icon,
-                        contentDescription = null, 
-                        tint = location.color, 
-                        modifier = Modifier.size(24.dp)
-                    )
-                }
-                Spacer(Modifier.width(12.dp))
-                Column(Modifier.weight(1f)) {
-                    @Suppress("DEPRECATION")
-                    Text(location.name, style = MaterialTheme.typography.titleMedium, fontWeight = FontWeight.ExtraBold, color = Color.Black, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                    if (location.description.isNotEmpty()) {
-                        @Suppress("DEPRECATION")
-                        Text(location.description, style = MaterialTheme.typography.bodySmall, color = Color.DarkGray, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                    }
-                }
-                IconButton(onClick = onClose, modifier = Modifier.size(24.dp)) { Icon(Icons.Default.Close, contentDescription = "Close", tint = Color.LightGray) }
-            }
-            Spacer(Modifier.height(12.dp))
-            Row(horizontalArrangement = Arrangement.Center) {
-                Surface(
-                    onClick = onDetailsClick,
-                    color = AppBackground,
-                    shape = RoundedCornerShape(10.dp),
-                    modifier = Modifier.height(36.dp).fillMaxWidth(),
-                    border = BorderStroke(1.dp, CoralRed.copy(alpha = 0.2f))
-                ) {
-                    Box(contentAlignment = Alignment.Center) {
-                        @Suppress("DEPRECATION")
-                        Text("Details", color = CoralRed, fontWeight = FontWeight.Bold, fontSize = 13.sp)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun UserLocationMarker(position: LatLng) {
-    MarkerComposable(state = rememberMarkerState(position = position)) {
-        Box(contentAlignment = Alignment.Center) {
-            Icon(Icons.Default.PersonPinCircle, contentDescription = null, tint = Color(0xFF1A73E8), modifier = Modifier.size(40.dp))
-            Box(modifier = Modifier.size(12.dp).offset(y = (-2).dp).background(Color.White, CircleShape))
-            Box(modifier = Modifier.size(8.dp).offset(y = (-2).dp).background(Color(0xFF1A73E8), CircleShape))
-        }
-    }
-    Circle(center = position, radius = 25.0, fillColor = Color(0xFF1A73E8).copy(alpha = 0.15f), strokeColor = Color(0xFF1A73E8).copy(alpha = 0.4f), strokeWidth = 2f)
-}
-
-@Composable
-fun MapOverlays(
-    hasLocationPermission: Boolean, 
-    @Suppress("UNUSED_PARAMETER") isLoadingLocation: Boolean,
-    isPinMode: Boolean,
-    isEditing: Boolean = false,
-    onLocationRequest: () -> Unit,
-    onTogglePinMode: () -> Unit,
-    onConfirmPin: () -> Unit,
-    onClearPin: () -> Unit,
-    tempPinSet: Boolean
-) {
-    Box(modifier = Modifier.fillMaxSize()) {
-        if (!isPinMode) {
-            Column(
-                modifier = Modifier.align(Alignment.BottomStart).padding(20.dp).padding(bottom = 90.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                if (hasLocationPermission) {
-                    FloatingActionButton(
-                        onClick = onLocationRequest,
-                        shape = CircleShape,
-                        containerColor = Color.White,
-                        contentColor = CoralRed,
-                        elevation = FloatingActionButtonDefaults.elevation(defaultElevation = 6.dp)
-                    ) {
-                        Icon(painter = painterResource(id = android.R.drawable.ic_menu_mylocation), contentDescription = "My Location", modifier = Modifier.size(24.dp))
-                    }
-                }
-
-                FloatingActionButton(
-                    onClick = onTogglePinMode,
-                    shape = CircleShape,
-                    containerColor = Color.White,
-                    contentColor = CoralRed,
-                    elevation = FloatingActionButtonDefaults.elevation(defaultElevation = 6.dp)
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.AddLocationAlt,
-                        contentDescription = "Custom Pin",
-                        modifier = Modifier.size(24.dp)
-                    )
-                }
-            }
-        }
-
-        if (isPinMode) {
-            // Header for pin mode
-            Surface(
-                modifier = Modifier.align(Alignment.TopCenter).padding(top = 16.dp),
-                shape = RoundedCornerShape(20.dp),
-                color = CoralRed,
-                shadowElevation = 8.dp
-            ) {
-                @Suppress("DEPRECATION")
-                Text(
-                    text = if (isEditing) "Move the pin or confirm location."
-                           else if (tempPinSet) "Pin placed! Confirm to continue."
-                           else "Tap on map to place your pin",
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
-                    color = Color.White,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-
-            // Bottom controls for pin mode
-            Column(
-                modifier = Modifier.align(Alignment.BottomCenter).padding(bottom = 40.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                if (tempPinSet) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(12.dp)
-                    ) {
-                        Button(
-                            onClick = onClearPin,
-                            colors = ButtonDefaults.buttonColors(containerColor = Color.White, contentColor = CoralRed),
-                            shape = RoundedCornerShape(50),
-                            border = BorderStroke(1.dp, CoralRed)
-                        ) {
-                            Icon(Icons.Default.Refresh, contentDescription = null, modifier = Modifier.size(18.dp))
-                            Spacer(Modifier.width(8.dp))
-                            @Suppress("DEPRECATION")
-                            Text("Change Location")
-                        }
-                        Button(
-                            onClick = onConfirmPin,
-                            colors = ButtonDefaults.buttonColors(containerColor = CoralRed),
-                            shape = RoundedCornerShape(50)
-                        ) {
-                            Icon(Icons.Default.Check, contentDescription = null, modifier = Modifier.size(18.dp))
-                            Spacer(Modifier.width(8.dp))
-                            @Suppress("DEPRECATION")
-                            Text("Confirm Location")
-                        }
-                    }
-                }
-
-                // Exit pin mode button
-                Surface(
-                    onClick = onTogglePinMode,
-                    color = Color.White.copy(alpha = 0.9f),
-                    shape = RoundedCornerShape(12.dp),
-                    border = BorderStroke(1.dp, Color.LightGray),
-                    modifier = Modifier.height(40.dp)
-                ) {
-                    Row(
-                        modifier = Modifier.padding(horizontal = 16.dp),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        Icon(Icons.Default.Close, contentDescription = null, tint = Color.DarkGray, modifier = Modifier.size(16.dp))
-                        Spacer(Modifier.width(8.dp))
-                        @Suppress("DEPRECATION")
-                        Text("Cancel", color = Color.DarkGray, fontWeight = FontWeight.Bold, fontSize = 14.sp)
-                    }
-                }
-            }
-        }
-
-        if (!hasLocationPermission && !isPinMode) PermissionCard(modifier = Modifier.align(Alignment.Center))
-    }
-}
-
-@Composable
-fun PermissionCard(modifier: Modifier) {
-    Card(
-        modifier = modifier.padding(32.dp),
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.errorContainer),
-        shape = RoundedCornerShape(16.dp)
-    ) {
-        @Suppress("DEPRECATION")
-        Text("Location permission is required to show your position on the map.", modifier = Modifier.padding(20.dp), color = MaterialTheme.colorScheme.onErrorContainer, fontWeight = FontWeight.Medium)
     }
 }
